docker/requirements.txt
torch==2.3.0
torchvision==0.18.0
wandb==0.17.0
pyparsing==3.1.0
biopython==1.79
matplotlib==3.8.3
matplotlib-inline==0.1.7
molgrid==0.5.3
numpy==1.23.5
pluggy==1.4.0
ProDy==2.4.1
rdkit==2023.9.6
scikit-image==0.23.2
scikit-learn==1.5.0